* gcc4 support * libw32dll update * Support for newer libquicktime
git-svn-id: svn://svn.code.sf.net/p/gmerlin/code/trunk@511 214d8ac5-8513-4f59-80f8-2f39b674a24a

+++ gmerlin_avdecoder/lib/libw32dll/wine/win32.c
@@ -2298,7 +2298,7 @@ static int WINAPI expLoadLibraryA(char* name)
     // we skip to the last backslash
     // this is effectively eliminating weird characters in
     // the text output windows
-//    fprintf(stderr, "expLoadLibraryA: %s\n", name);
+
     lastbc = strrchr(name, '\\');
     if (lastbc)
     {
@@ -2623,7 +2623,7 @@ static int WINAPI expGetWindowThreadProcessId(HWND win, int *pid_data)
     dbgprintf("GetWindowThreadProcessId(0x%x, 0x%x) => %d\n",
 	win, pid_data, tid);
     if (pid_data)
-	(int)*pid_data = tid;
+	*pid_data = tid;
     return tid;
 }
